+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 - Add a hinge-angle based bending force to VBDIntegrator.
 - Add the `Device.sm_count` property to get the number of streaming multiprocessors on a CUDA device
   ([GH-584](https://github.com/NVIDIA/warp/issues/584)).
+- Add an example to show mesh sampling using a CDF
+  ([GH-476](https://github.com/NVIDIA/warp/issues/476)).
 
 ### Changed